Expert Functions

The LRMoE package supports a collection of distributions commonly used for modelling insurance claim frequency and severity.

Type Hierarchy

Expert functions are implemented as subtypes of the AnyExpert type in this package. A good number of expert functions are simply wrappers around the UnivariateDistribution type in Distributions.jl (details here), and functions such as pdf and cdf are also directly using those in Distributions.jl.

Specific to actuarial application, a collection of commonly used distributions are also included in this package, e.g. Burr, GammaCount, etc. See below for details.

Expert functions can be defined either on the real line, or only on nonnegative values (as is usually the case for actuarial loss modelling).

# Abstract type: support of expert function
abstract type ExpertSupport end
struct RealValued <: ExpertSupport end
struct NonNegative <: ExpertSupport end

Considering zero inflation is prominant in many actuarial applications, expert functions can be either zero-inflated or not, provided that they are supported only on nonnegative values.

# Abstract type: whether the expert is zero-inflated
abstract type ZeroInflation end
struct ZI <: ZeroInflation end
struct NonZI <: ZeroInflation end

Each expert function is a univariate distribution (see here for UnivariateDistribution in Distributions.jl), with appropriate indication of support and zero inflation.

# Abstract type: AnyExpert
abstract type AnyExpert{s<:ExpertSupport, z<:ZeroInflation, d<:UnivariateDistribution} end
# Discrete or Continuous
const DiscreteExpert{s<:ExpertSupport, z<:ZeroInflation} = 
                AnyExpert{s, z, DiscreteUnivariateDistribution}
const ContinuousExpert{s<:ExpertSupport, z<:ZeroInflation} = 
                AnyExpert{s, z, ContinuousUnivariateDistribution}
# Real-valued expert distributions
const RealDiscreteExpert = DiscreteExpert{RealValued, NonZI}
const RealContinuousExpert = ContinuousExpert{RealValued, NonZI}
# Nonnegative-valued expert distributions
const NonNegDiscreteExpert{z<:ZeroInflation} = DiscreteExpert{NonNegative, z}
const NonNegContinuousExpert{z<:ZeroInflation} = ContinuousExpert{NonNegative, z}
# Zero-inflated
const ZIDiscreteExpert = NonNegDiscreteExpert{ZI}
const ZIContinuousExpert = NonNegContinuousExpert{ZI}
# Non zero-inflated
const NonZIDiscreteExpert = NonNegDiscreteExpert{NonZI}
const NonZIContinuousExpert = NonNegContinuousExpert{NonZI}

Discrete Distributions (Frequency Modelling)

LRMoE.BinomialExpert — Type

BinomialExpert(n, p)

PMF:

\[P(X = k) = {n \choose k}p^k(1-p)^{n-k}, \quad \text{ for } k = 0,1,2, \ldots, n.\]

See also: Binomial Distribution (Wikipedia)

LRMoE.NegativeBinomialExpert — Type

NegativeBinomialExpert(n, p)

PMF:

\[P(X = k) = \frac{\Gamma(k+r)}{k! \Gamma(r)} p^r (1 - p)^k, \quad \text{for } k = 0,1,2,\ldots.\]

See also: Negative Binomial Distribution (Wolfram)

LRMoE.PoissonExpert — Type

PoissonExpert(λ)

PMF:

\[P(X = k) = \frac{\lambda^k}{k!} e^{-\lambda}, \quad \text{ for } k = 0,1,2,\ldots.\]

See also: Poisson Distribution (Wikipedia)

Continuous Distributions (Severity Modelling)

LRMoE.BurrExpert — Type

BurrExpert(k, c, λ)

PDF:

\[f(x; k, c, \lambda) = \frac{kc}{\lambda} \left( \frac{x}{\lambda} \right)^{c-1} \left( 1+ \left( \frac{x}{\lambda} \right)^{c} \right)^{-k-1}, \quad x \geq 0\]

See also: Burr Distribution (Mathworks, implemented in this package), Burr Distribution (Wikipedia, with λ = 1)

LRMoE.GammaExpert — Type

GammaExpert(k, θ)

PDF:

\[f(x; k, \theta) = \frac{x^{k-1} e^{-x/\theta}}{\Gamma(k) \theta^k}, \quad x > 0\]

See also: Gamma Distribution (Wikipedia), shape-scale parameterization

LRMoE.InverseGaussianExpert — Type

InverseGaussianExpert(μ, λ)

PDF:

\[f(x; \mu, \lambda) = \sqrt{\frac{\lambda}{2\pi x^3}} \exp\left(\frac{-\lambda(x-\mu)^2}{2\mu^2x}\right), \quad x > 0\]

See also: Inverse Gaussian Distribution (Wikipedia)

LRMoE.LogNormalExpert — Type

LogNormalExpert(μ, σ)

PDF:

\[f(x; \mu, \sigma) = \frac{1}{x \sqrt{2 \pi \sigma^2}} \exp \left( - \frac{(\log(x) - \mu)^2}{2 \sigma^2} \right), \quad x > 0\]

See also: Lognormal Distribution (Wikipedia)

LRMoE.WeibullExpert — Type

WeibullExpert(k, θ)

PDF:

\[f(x; k, \theta) = \frac{k}{\theta} \left( \frac{x}{\theta} \right)^{k-1} e^{-(x/\theta)^k}, \quad x \geq 0\]

See also: Weibull Distribution (Wikipedia)

Zero Inflation

Zero inflation is supported for all discrete and continuous experts. They can be constructed by adding ZI in front of an expert function, with an additional parameter p (or p0 if the expert already uses p, e.g. binomial) for modelling a probability mass at zero. Zero-inflated experts are used in the same way as their non-zero-inflated counterpart. A complete list of zero-inflated expert functions is given below.

ZIBinomialExpert(p0, n, p)
ZINegativeBinomialExpert(p0, n, p)
ZIPoissonExpert(p, λ)

ZIBurrExpert(p, k, c, λ)
ZIGammaExpert(p, k, θ)
ZIInverseGaussianExpert(p, μ, λ)
ZILogNormalExpert(p, μ, σ)
ZIWeibullExpert(p, k, θ)